+++ packages/@storefront/core/src/core/tag/index.ts
@@ -17,6 +17,7 @@ class Tag<P extends object = any, S extends object = any, A extends object = any
   _provides: Record<string, ((props: P & Tag.Props, state: S) => (aliases: A) => void)> = {};
   _consumes: string[] = [];
   _eventHandlers: [string, () => void][] = [];
+  _lookups: [string[], () => void][] = [];
   isInitialized: boolean = false;
   props: P & Tag.Props = <any>{};
   state: S = <any>{};
@@ -39,6 +40,7 @@ class Tag<P extends object = any, S extends object = any, A extends object = any
 
   subscribe<T>(event: string, handler: (data?: T) => void) {
     this.flux.on(event, handler);
+
     if (this._eventHandlers.length === 0) {
       this.one(Phase.UNMOUNT, this._removeEventHandlers);
     }
@@ -48,6 +50,16 @@ class Tag<P extends object = any, S extends object = any, A extends object = any
 
   subscribeOnce(event: string, handler: (event: string, data?: any) => void) {
     this.flux.once(event, handler);
+  }
+
+  subscribeWith<T>(events: string[], handler: (data?: T) => void) {
+    this.flux.all(events, handler);
+
+    if (this._lookups.length === 0) {
+      this.one(Phase.UNMOUNT, this._removeLookups);
+    }
+
+    this._lookups.push([events, handler as any]);
   }
 
   provide(alias: string, resolve: (props: P, state: S, aliases: A) => void = (_, state) => state) {
@@ -73,6 +85,8 @@ class Tag<P extends object = any, S extends object = any, A extends object = any
   }
 
   _removeEventHandlers = () => this._eventHandlers.forEach(([event, handler]) => this.flux.off(event, handler));
+
+  _removeLookups = () => this._lookups.forEach(([events, handler]) => this.flux.allOff(events, handler));
 }
 
 interface Tag<P extends object, S extends object, A extends object>