+++ packages/@storefront/core/CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [minor]
+### Added
+- SF-1170: Add support for `subscribeWith()` method.
+  - Allows components to invoke callbacks after each event in a given collection has been emitted at least once.
+  - Automatically unregisters callbacks when the component is unmounted.
+
 ## [1.51.5] - 2018-11-14
 ### Fixed
 - SF-1132: Default to an empty array for array methods and pass through the rest of the request.